Move generic free energy functions to free_energy submodule
* move test_pair_overlap_from_ukln to test_bar.py
