+++ timemachine/fe/free_energy.py
@@ -4,7 +4,11 @@ from typing import Iterable, List, Optional, Sequence, Tuple, Union, overload
 import numpy as np
 from numpy.typing import NDArray
 
+from timemachine.constants import BOLTZ
 from timemachine.fe import model_utils, topology
+from timemachine.fe.bar import bar_with_bootstrapped_uncertainty, df_err_from_ukln, pair_overlap_from_ukln
+from timemachine.fe.energy_decomposition import EnergyDecomposedState, compute_energy_decomposed_u_kln, get_batch_u_fns
+from timemachine.fe.plots import make_dG_errs_figure, make_overlap_detail_figure, make_overlap_summary_figure
 from timemachine.fe.stored_arrays import StoredArrays
 from timemachine.fe.utils import get_mol_masses, get_romol_conf
 from timemachine.ff import ForcefieldParams
@@ -312,3 +316,186 @@ def sample(initial_state: InitialState, md_params: MDParams, max_buffer_frames: 
     assert np.all(np.isfinite(all_coords[-1])), "Production resulted in a nan"
 
     return all_coords, all_boxes
+
+
+def estimate_free_energy_given_initial_states(
+    initial_states,
+    md_params,
+    temperature,
+    prefix,
+    keep_idxs,
+):
+    """
+    Estimate free energies given pre-generated samples. This implements the pair-BAR method, where
+    windows assumed to be ordered with good overlap, with the final free energy being a sum
+    of the components. The constants below are:
+
+    L: the number of lambda windows
+    T: the number of samples
+    N: the number of atoms
+    P: the number of components in the energy function.
+
+    Parameters
+    ----------
+    initial_states: list of InitialState
+        Initial state objects
+
+    md_params: MDParams
+        Detailing specifics of each simulation
+
+    temperature: float
+        Temperature the system was run at
+
+    prefix: str
+        A prefix that we append to the BAR overlap figures
+
+    keep_idxs: list of int
+        Which states we keep samples for. Must be positive.
+
+    Return
+    ------
+    SimulationResult
+        object containing results of the simulation
+
+    """
+
+    # run n_lambdas simulations in sequence
+    u_kln_by_component_by_lambda, stored_frames, stored_boxes = run_sequential_sims_given_initial_states(
+        initial_states, md_params, temperature, keep_idxs
+    )
+
+    # "pair BAR" free energy analysis
+    kBT = BOLTZ * temperature
+    beta = 1 / kBT
+
+    all_dGs = []
+    all_errs = []
+    for lamb_idx, u_kln_by_component in enumerate(u_kln_by_component_by_lambda):
+        # pair BAR
+        u_kln = u_kln_by_component.sum(0)
+
+        w_fwd = u_kln[1, 0] - u_kln[0, 0]
+        w_rev = u_kln[0, 1] - u_kln[1, 1]
+
+        df, df_err = bar_with_bootstrapped_uncertainty(w_fwd, w_rev)  # reduced units
+        dG, dG_err = df / beta, df_err / beta  # kJ/mol
+
+        message = f"{prefix} BAR: lambda {lamb_idx} -> {lamb_idx + 1} dG: {dG:.3f} +- {dG_err:.3f} kJ/mol"
+        print(message, flush=True)
+
+        all_dGs.append(dG)
+        all_errs.append(dG_err)
+
+    # (energy components, lambdas, energy fxns = 2, sampled states = 2, frames)
+    ukln_by_lambda_by_component = np.array(u_kln_by_component_by_lambda).swapaxes(0, 1)
+
+    # compute more diagnostics
+    overlaps_by_lambda = [pair_overlap_from_ukln(u_kln) for u_kln in ukln_by_lambda_by_component.sum(axis=0)]
+
+    dG_errs_by_lambda_by_component = np.array(
+        [[df_err_from_ukln(u_kln) / beta for u_kln in ukln_by_lambda] for ukln_by_lambda in ukln_by_lambda_by_component]
+    )
+    overlaps_by_lambda_by_component = np.array(
+        [[pair_overlap_from_ukln(u_kln) for u_kln in ukln_by_lambda] for ukln_by_lambda in ukln_by_lambda_by_component]
+    )
+
+    # generate figures
+    U_names = [type(U_fn).__name__ for U_fn in initial_states[0].potentials]
+    lambdas = [s.lamb for s in initial_states]
+
+    overlap_detail_png = make_overlap_detail_figure(
+        U_names,
+        all_dGs,
+        all_errs,
+        u_kln_by_component_by_lambda,
+        temperature,
+        prefix,
+    )
+    dG_errs_png = make_dG_errs_figure(U_names, lambdas, all_errs, dG_errs_by_lambda_by_component)
+    overlap_summary_png = make_overlap_summary_figure(
+        U_names, lambdas, overlaps_by_lambda, overlaps_by_lambda_by_component
+    )
+
+    return SimulationResult(
+        all_dGs,
+        all_errs,
+        dG_errs_by_lambda_by_component,
+        overlaps_by_lambda,
+        overlaps_by_lambda_by_component,
+        dG_errs_png,
+        overlap_summary_png,
+        overlap_detail_png,
+        stored_frames,
+        stored_boxes,
+        initial_states,
+        md_params,
+    )
+
+
+def run_sequential_sims_given_initial_states(
+    initial_states,
+    md_params,
+    temperature,
+    keep_idxs,
+):
+    """Sequentially run simulations at each state in initial_states,
+    returning summaries that can be used for pair BAR, energy decomposition, and other diagnostics
+
+    Returns
+    -------
+    decomposed_u_klns: [n_lams - 1, n_components, 2, 2, n_frames] array
+    stored_frames: coord trajectories, one for each state in keep_idxs
+    stored_boxes: box trajectories, one for each state in keep_idxs
+
+    Notes
+    -----
+    * Memory complexity:
+        Memory demand should be no more than that of 2 states worth of frames.
+
+        Requesting too many states in keep_idxs may blow this up,
+        so best to keep to first and last states in keep_idxs.
+
+        This restriction may need to be relaxed in the future if:
+        * We decide to use MBAR(states) rather than sum_i BAR(states[i], states[i+1])
+        * We use online protocol optimization approaches that require more states to be kept on-hand
+    """
+    stored_frames = []
+    stored_boxes = []
+
+    tmp_states = [None] * len(initial_states)
+
+    # u_kln matrix (2, 2, n_frames) for each pair of adjacent lambda windows and energy term
+    u_kln_by_component_by_lambda = []
+
+    keep_idxs = keep_idxs or []
+    if keep_idxs:
+        assert all(np.array(keep_idxs) >= 0)
+
+    for lamb_idx, initial_state in enumerate(initial_states):
+
+        # run simulation
+        cur_frames, cur_boxes = sample(initial_state, md_params)
+        print(f"completed simulation at lambda={initial_state.lamb}!")
+
+        # keep samples from any requested states in memory
+        if lamb_idx in keep_idxs:
+            stored_frames.append(cur_frames)
+            stored_boxes.append(cur_boxes)
+
+        # construct EnergyDecomposedState for current lamb_idx,
+        # but keep no more than 2 of these states in memory at once
+        if lamb_idx >= 2:
+            tmp_states[lamb_idx - 2] = None
+
+        bound_impls = [p.bound_impl(np.float32) for p in initial_state.potentials]
+        cur_batch_U_fns = get_batch_u_fns(bound_impls, temperature)
+
+        tmp_states[lamb_idx] = EnergyDecomposedState(cur_frames, cur_boxes, cur_batch_U_fns)
+
+        # analysis that depends on current and previous state
+        if lamb_idx > 0:
+            state_pair = [tmp_states[lamb_idx - 1], tmp_states[lamb_idx]]
+            u_kln_by_component = compute_energy_decomposed_u_kln(state_pair)
+            u_kln_by_component_by_lambda.append(u_kln_by_component)
+
+    return np.array(u_kln_by_component_by_lambda), stored_frames, stored_boxes
